extract match drag-n-drop code and include it in level groups
