+++ apps/src/sites/studio/pages/levels/_level_group.js
@@ -7,6 +7,7 @@ import getScriptData from '@cdo/apps/util/getScriptData';
 import * as codeStudioLevels from '@cdo/apps/code-studio/levels/codeStudioLevels';
 import {SingleLevelGroupDialog} from '@cdo/apps/lib/ui/LegacyDialogContents';
 import i18n from '@cdo/locale';
+import {initMatch} from '@cdo/apps/code-studio/levels/match';
 
 window.Multi = require('@cdo/apps/code-studio/levels/multi.js');
 window.TextMatch = require('@cdo/apps/code-studio/levels/textMatch.js');
@@ -25,6 +26,8 @@ $(document).ready(() => {
       initData.last_attempt
     );
   }
+
+  initMatch();
 });
 
 function initLevelGroup(levelCount, currentPage, lastAttempt) {
